requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
sentence-transformers
faiss-cpu
pyttsx3
streamlit
PyMuPDF
